feat(#8): Job/Project redesign notes
