+++ src/main/java/git/tracehub/agents/github/GhJob.java
@@ -1,0 +1,55 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2023-2024 Tracehub.git
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package git.tracehub.agents.github;
+
+import com.jcabi.github.Repo;
+import git.tracehub.Job;
+import lombok.RequiredArgsConstructor;
+
+/**
+ * Job in GitHub.
+ *
+ * @since 0.0.0
+ */
+@RequiredArgsConstructor
+public final class GhJob implements Job {
+
+    /**
+     * Repo.
+     */
+    private final Repo repo;
+
+    /**
+     * Job name.
+     */
+    private final String name;
+
+    @Override
+    public String asString() throws Exception {
+        return new GhContent(
+            this.repo,
+            ".trace/jobs/%s".formatted(this.name)
+        ).asString();
+    }
+}
